try to make page nav for modules
