+++ apps/web/components/user-profile/modules/index.tsx
@@ -1,50 +1,16 @@
-import { ModuleSimple } from 'types'
-import { ModuleStatus } from '@modtree/types'
-import { ModuleListSection } from '@/ui/settings'
-
-type ModuleContent = Record<ModuleStatus, ModuleSimple[]>
-
-const moduleContent: ModuleContent = {
-  done: [
-    {
-      title: 'Linear Algebra',
-      moduleCode: 'MA2001',
-    },
-    {
-      title: 'Calculus',
-      moduleCode: 'MA2002',
-    },
-  ],
-  doing: [
-    {
-      title: 'Introduction to Geometry',
-      moduleCode: 'MA2219',
-    },
-    {
-      title: 'Design Thinking',
-      moduleCode: 'DTK1234',
-    },
-  ],
-  notTaken: [],
-}
+import { Pages } from 'types'
+import { Main } from './main'
+import { useState } from 'react'
 
 export function ModulesTabContent() {
-  return (
-    <>
-      <ModuleListSection
-        contents={moduleContent.doing}
-        title="Modules Doing"
-        addButtonText="Add doing"
-        summary="This is a list of modules that are currently in progess this semester. They will be automatically marked as when finals week is over."
-        emptySummary="There are no modules that are currently in progress."
-      />
-      <ModuleListSection
-        contents={moduleContent.done}
-        title="Modules Done"
-        addButtonText="Add done"
-        summary="This is a list of completed modules."
-        emptySummary="There are no modules that are completed."
-      />
-    </>
-  )
+  const [page, setPage] = useState<Pages['Modules']>('main')
+  if (page === 'main') {
+    return <Main setPage={setPage} />
+  }
+  if (page === 'add-doing') {
+    return <Main setPage={setPage} />
+  }
+  if (page === 'add-done') {
+    return <Main setPage={setPage} />
+  }
 }
