+++ apps/web/pages/index.tsx
@@ -1,4 +1,4 @@
-import { FloatingUserButton } from '@/components/buttons'
+import { FloatingActionButton, FloatingUserButton } from '@/components/buttons'
 import { FullScreenOverlay } from '@/components/Views'
 import ModtreeFlow from '@/flow'
 import { HomeLoader } from '@/components/Loader'
@@ -36,6 +36,7 @@ export default function Modtree() {
           <RootSearchBox />
           <FullScreenOverlay>
             <FloatingUserButton />
+            <FloatingActionButton />
           </FullScreenOverlay>
           <UserProfileModal />
           <DebugModal />
