+++ apps/web/types/index.d.ts
@@ -86,4 +86,5 @@ export type DegreeSummary = {
 
 export type Pages = {
   Degrees: 'main' | 'add-new'
+  Modules: 'main' | 'add-doing' | 'add-done'
 }
