+++ apps/web/ui/settings/lists/modules.tsx
@@ -9,17 +9,21 @@ export function ModuleListSection(props: {
   summary: string
   emptySummary: string
   addButtonText: string
+  onAddClick: () => void
 }) {
   const hasModules = props.contents.length !== 0
-  const { title, contents, summary, emptySummary, addButtonText } = props
   return (
     <div className="mb-12">
-      <SettingsSection title={title} addButtonText={addButtonText}>
+      <SettingsSection
+        title={props.title}
+        addButtonText={props.addButtonText}
+        onAddClick={props.onAddClick}
+      >
         {hasModules ? (
           <>
-            <p>{summary}</p>
+            <p>{props.summary}</p>
             <div className="ui-rectangle flex flex-col overflow-hidden">
-              {contents.map((module, index) => (
+              {props.contents.map((module, index) => (
                 <Row.Module key={dashed(module.moduleCode, index)}>
                   <span className="font-semibold">{module.moduleCode}</span>
                   <span className="mx-1">/</span>
@@ -29,7 +33,7 @@ export function ModuleListSection(props: {
             </div>
           </>
         ) : (
-          <p>{emptySummary}</p>
+          <p>{props.emptySummary}</p>
         )}
       </SettingsSection>
     </div>
